% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.algn.foot.rng.R
\name{def.algn.foot.rng}
\alias{def.algn.foot.rng}
\title{Align Footprint Map with Remote Sensing Data}
\usage{
def.algn.foot.rng(PHIcr_pad, nortF, eastF, RS)
}
\arguments{
\item{PHIcr_pad}{Matrix representing the footprint map.}

\item{nortF}{Numeric vector representing the north extent of the footprint.}

\item{eastF}{Numeric vector representing the east extent of the footprint.}

\item{RS}{List containing 'nort' and 'east' vectors for remote sensing data extents.}
}
\value{
List containing the adjusted footprint map matrix and the updated extents (nortF and eastF).
}
\description{
Function definition. This function adjusts the dimensions of a footprint map
(PHIcr_pad) to match and align with the dimensions of remote sensing (RS) data. It handles both cases
where the footprint map is smaller or larger than the RS data extent in all directions.
}
\examples{
# Example usage:
# result <- alignFootprintMap(PHIcr_pad, nortF, eastF, RS)
# aligned_PHIcr_pad <- result$PHIcr_pad
# updated_nortF <- result$nortF
# updated_eastF <- result$eastF

}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none.
}
\author{
Stefan Metzger \email{smetzger@atmofacts.com}
Andrei Serafimovich
}
\keyword{flux}
\keyword{mapping}
\keyword{visualization,}
