% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.bbox.R
\name{def.bbox}
\alias{def.bbox}
\title{definition function to create bounding box}
\usage{
def.bbox(
<<<<<<< HEAD
  site = c("NEON:CPER", "CHEESEHEAD")[2],
=======
  DirInp = NULL,
>>>>>>> atmo/maps
  crs = "EPSG:4326",
  pad_box = FALSE,
  pad_degree = 0.5,
  lat = NULL,
  lon = NULL
)
}
\arguments{
<<<<<<< HEAD
\item{site}{the site to use to grab the bounding box}

=======
>>>>>>> atmo/maps
\item{crs}{coordinate reference system (Defaults to "EPSG:4326")}

\item{pad_box}{logical to determine if padding should be applied to bounding box (defaults to FALSE)}

\item{pad_degree}{decimal degree of latitude and longitude to pad the bounding box (defaults to 0.1)}
<<<<<<< HEAD
=======

\item{dir}{directory where shapefiles are located}
>>>>>>> atmo/maps
}
\value{
list containing shapefile data and bbox
}
\description{
Function definition. Read in shapefile or create a bounding box
by adding an extent to Lat/Lon for grabbing STAC data with a padding option.
}
\examples{
Currently none
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
David Durden
<<<<<<< HEAD
}
\keyword{constants}
\keyword{natural}
=======
Stefan Metzger \email{smetzger@atmofacts.com}
}
\keyword{Currently}
\keyword{none}
>>>>>>> atmo/maps
