% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.plot.flux.animate.R
\name{def.plot.flux.animate}
\alias{def.plot.flux.animate}
\title{Definition function: Plot fluxes on an animated gif.}
\usage{
def.plot.flux.animate(
  input_folder,
  output_file,
  nodata_value = 0,
  colormap = "YlOrRd",
  basemap_style = "osm",
  palette_style = "equal",
  alpha = 0.4,
  legend = TRUE,
  delay = 1
)
}
\arguments{
\item{input_folder}{A folder of geotiff files.}

\item{output_file}{The path of the output file (.gif)}

\item{nodata_value}{Nodata value from flux data}

\item{colormap}{palette_explorer() for palette options}

\item{palette_style}{quantile, equal, cont, cat}

\item{alpha}{Raster opacity}

\item{legend}{Boolean. If legend plots on the map}

\item{delay}{The delay time in the animation.}
}
\description{
Function definition. This function creates an animated flux map visualization for temporal flux map data contained in a folder.
}
\examples{

def.plot.flux.animation(
input_folder = "/path/to/input/folder",  # Replace with your input folder path
output_file = "/path/to/output/animation.gif",  # Replace with your desired output GIF path
nodata_value = 0,
colormap = 'YlOrRd',
basemap_style = 'osm',
palette_style = 'equal',
alpha = 0.4,
legend = TRUE
)
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none.
}
\author{
Sam Bower \email{sbower@atmofacts.com}
}
\keyword{animation}
\keyword{flux}
\keyword{mapping,}
\keyword{visualization,}
