% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.plot.flux.interactive.R
\name{def.plot.flux.interactive}
\alias{def.plot.flux.interactive}
\title{Definition function: Plot fluxes on an interactive map.}
\usage{
def.plot.flux.interactive(
  input_path,
  nodata_value = 0,
  alpha = 0.7,
  colormap = "YlOrRd",
  save_path = NULL
)
}
\arguments{
\item{input_path}{A geotiff file or folder of geotiff files to plot on the basemap.}

\item{nodata_value}{Nodata value from flux data}

\item{alpha}{Raster opacity}

\item{colormap}{palette_explorer() for palette options}
}
\description{
Function definition. This function creates an interactive flux map visualization that can be used in R viewer or in a web browser.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none.
}
\author{
Sam Bower \email{sbower@atmofacts.com}
}
\keyword{flux}
\keyword{mapping}
\keyword{visualization,}
