% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.plot.flux.static.R
\name{def.plot.flux.static}
\alias{def.plot.flux.static}
\title{Definition function: Plot fluxes on a static basemap}
\usage{
def.plot.flux.static(
  inputPath,
  outputPath,
  basemap_style = "bing",
  alpha = 0.5,
  nodata_value = 0,
  colormap = "YlOrRd",
  style = "equal",
  color_n = 7,
  use_basemap = FALSE,
  background_color = "white",
  units = "Units",
  legend_bg_alpha = 1
)
}
\arguments{
\item{inputPath}{A spatial file or folder of spatial files to plot on the basemap.}

\item{outputPath}{A string for the folder location to save output maps.}

\item{basemap_style}{Basemap styles in OpenStreetMap (more in openmap())}

\item{alpha}{Raster opacity}

\item{nodata_value}{The nodata value specified from the flux data.}

\item{colormap}{palette_explorer() for palette options}

\item{style}{quantile, equal, cont, cat}

\item{color_n}{color categories in palette}
}
\description{
Function definition. This function creates a flux map visualization on a user defined basemap.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none.
}
\author{
Sam Bower \email{sbower@atmofacts.com}
}
\keyword{flux}
\keyword{mapping}
\keyword{visualization,}
