% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spat.data.cube.R
\name{def.spat.data.cube}
\alias{def.spat.data.cube}
\title{definition function to download data from Microsoft planetary computer}
\usage{
def.spat.data.cube(
  start_date,
  end_date,
  box,
  collection = c("modis-15A2H-061", "sentinel-3-slstr-lst-l2-netcdf")[2],
  asset_name = c("Lai_500m", "lst-in")[2],
  srs = "EPSG:4326",
  dx = 0.1,
  dy = 0.1,
  dt = "P1M",
  aggregation = "mean",
  resampling = "near"
)
}
\arguments{
\item{start_date}{start date as character format yyyy-mm-dd}

\item{end_date}{end date as character format yyyy-mm-dd}

\item{box}{numberic vector in the format of (xmin, ymin, xmax, ymax)}

\item{collection}{name of planetary collection}

\item{asset_name}{mame of asset}

\item{srs}{target spatial reference system as a string; can be a proj4 definition, WKT, or in the form "EPSG:XXXX"}

\item{dx}{size of pixels in x-direction (longitude / easting)}

\item{dy}{size of pixels in y-direction (latitude / northing)}

\item{dt}{size of pixels in time-direction, expressed as ISO8601 period string (only 1 number and unit is allowed) such as "P16D"}

\item{aggregation}{aggregation method as string, defining how to deal with pixels containing data from multiple images, can be "min", "max", "mean", "median", or "first"}

\item{resampling}{resampling method used in gdalwarp when images are read, can be "near", "bilinear", "bicubic" or others as supported by gdalwarp (see https://gdal.org/programs/gdalwarp.html)}
}
\value{
A data cube proxy object
}
\description{
Function definition. Download data from Microsoft planetary computer
}
\examples{
ingest_planetary_data(start_date = "2022-01-01", end_date = "2023-07-01", box =  c("xmin" = -123, "ymin" = 39, "xmax" = -122, "ymax" = 40))
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
David Durden
Stefan Metzger \email{smetzger@atmofacts.com}
}
\keyword{Currently}
\keyword{none}
