% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.spatialize.flux.foot.R
\name{def.spatialize.flux.foot}
\alias{def.spatialize.flux.foot}
\title{Spatialize Flux Footprint Data into Raster Format}
\usage{
def.spatialize.flux.foot(
  flux_input,
  lower_left_X,
  lower_left_Y,
  nodata_value = 0,
  cell_size = 10,
  crs = "+proj=utm +zone=13 +datum=WGS84 +units=m +no_defs",
  write = FALSE,
  output_filename = NULL,
  file_format = "GTiff"
)
}
\arguments{
\item{flux_input}{The input flux data, which can be a path to an ASCII grid or CSV file, a matrix,
or a data frame.}

\item{lower_left_X}{The X coordinate of the lower left corner of the output raster.}

\item{lower_left_Y}{The Y coordinate of the lower left corner of the output raster.}

\item{nodata_value}{The value to be considered as 'no data' in the input flux data. Defaults to 0.}

\item{cell_size}{The size of each cell in the output raster, assumed to be square. Defaults to 10.}

\item{crs}{The coordinate reference system for the output raster, specified in PROJ.4 format.
Defaults to "+proj=utm +zone=13 +datum=WGS84 +units=m +no_defs".}

\item{write}{Logical, whether to write the output raster to a file. Defaults to FALSE.}

\item{output_filename}{The filename (including path) for the output file if write is TRUE.
If NULL and write is TRUE, an error is thrown.}

\item{file_format}{The format of the output file, defaults to 'GTiff'.}
}
\value{
A raster object representing the spatialized flux data.
}
\description{
This function processes flux footprint data, spatializing it into a raster format
suitable for GIS analysis. It supports reading flux data from various formats, setting the
spatial extent, and defining the coordinate reference system (CRS). The output can optionally
be written to a file in GeoTiff format.
}
\examples{
# Example usage:
raster <- def.spatialize.flux.foot(flux_input = "path/to/data.csv",
                                   lower_left_X = 100000,
                                   lower_left_Y = 400000,
                                   cell_size = 10,
                                   crs = "+proj=utm +zone=13 +datum=WGS84 +units=m +no_defs",
                                   write = TRUE,
                                   output_filename = "output.tif",
                                   file_format = 'GTiff')

}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
\code{\link[raster]{writeRaster}}, \code{\link[raster]{raster}}, \code{\link[sp]{CRS}}
}
\author{
Sam Bower \email{sbower@atmofacts.com}
}
\keyword{GIS,}
\keyword{flux}
\keyword{raster,}
\keyword{spatial,}
