% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.plot.flux.spatial.R
\name{plot.flux.spatial}
\alias{plot.flux.spatial}
\title{Definition function: Plot fluxes on a basemap}
\usage{
\method{plot}{flux.spatial}(
  input_path,
  output_path,
  colormap = "viridis",
  opacity = 0.5,
  basemap_style = "OpenStreetMap",
  interactive = FALSE
)
}
\arguments{
\item{input_path}{A spatial file or folder of spatial files to plot on the basemap.}

\item{output_path}{A string for the folder location to save output maps.}

\item{basemap_style}{For a list of basemap styles use names(providers)}

\item{interactive}{If FALSE - plots the flux(es) onto a static basemap. If TRUE, plots the flux(es) onto an interactive basemap.}
}
\value{

}
\description{
Function definition. This function saves a GeoTiff flux map onto a static basemap or an interactive html map.
}
\examples{

path_in = "/path/to/folder/of/tif/files"
path_out = "/path/to/desired/output/location"
plot.flux.spatial(path_in, path_out, 'viridis', opacity = 0.5,'OpenStreetMap', TRUE)

}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none.
}
\author{
Sam Bower \email{sbower@atmofacts.com}
}
\keyword{flux}
\keyword{mapping}
\keyword{visualization,}
