% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.spat.data.R
\name{wrap.spat.data.grab}
\alias{wrap.spat.data.grab}
\title{wrapper function to generate spatial targets}
\usage{
wrap.spat.data.grab(
  dir = tempdir(),
  dateBgn = "2019-05-01",
  dateEnd = "2019-11-30",
  site = c("CPER", "CHEESEHEAD")[2],
  dt = "P1M",
  dx = 0.1,
  dy = 0.1
)
}
\arguments{
\item{dir}{A directory of .tif files for scoring}

\item{site_id}{The side ID of the area being forecasted. Currently "august_complex" is the only option.}
}
\value{
message from minio submission
}
\description{
Wrapper definition. Generate spatial targets
}
\examples{
spat4cast_submit(dir = "targets", site_id = "august_complex")
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007
}
\seealso{
Currently none
}
\author{
David Durden
Stefan Metzger \email{smetzger@atmofacts.com}
}
\keyword{Currently}
\keyword{none}
